+++ views/admin_twig/de/module_options.php
@@ -9,6 +9,15 @@ declare(strict_types=1);
 
 $aLang = [
     'charset' => 'UTF-8',
+    'tbcluser_greetings' => 'Greetings',
+
+    'OEMODULETEMPLATE_GREETING_TITLE' => 'Beispiel Admin Controller',
+    'OEMODULETEMPLATE_GREETING_MESSAGE_TEXT' => 'Begrüßungsnachricht: ',
+    'OEMODULETEMPLATE_NO_GREETING_TEXT' => 'Es wurde keine Begrüßungsnachricht hinzugefügt!',
+    'OEMODULETEMPLATE_ALLOW_GREETING' => 'Benutzer erlauben die Begrüßungsnachricht zu setzen',
+    'OEMODULETEMPLATE_HELP_ALLOW_GREETING' => 'Dies ist ein Beispiel, wie ein Admintemplate erweitert werden kann. Momentan ist keine Funktionalität hinter dieser Checkbox hinterlegt, es wird keine Einstellung in der Datenbank gespeichert.',
+
+    # Module settings
     'SHOP_MODULE_GROUP_oemoduletemplate_main' => 'Einstellungen',
     'SHOP_MODULE_oemoduletemplate_GreetingMode' => 'Begrüßungsmodus',
     'SHOP_MODULE_oemoduletemplate_GreetingMode_generic' => 'höflich',